condaenv.8xxhbaxa.requirements.txt
torchviz
nbmerge
gym
atari-py
box2d-py
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.3.0/de_core_news_sm-2.3.0.tar.gz